Correction

+++ 0x0C-web_server/7-puppet_install_nginx_web_server.pp
@@ -8,8 +8,10 @@ package { 'nginx':
     ensure => 'installed',
 }
 exec { 'nginx root':
-    command => 'echo "Hello World!" | sudo tee /var/www/html/index.nginx-debian.html',
-    path    => '/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin',
+    command  => 'echo "Hello World!" | sudo tee /var/www/html/index.nginx-debian.html',
+    path     => '/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin',
+    escalate => true,
+    unless   => 'test -e /var/www/html/index.nginx-debian.html',
 }
 file { '/var/www/html/index.nginx-debian.html':
     ensure  => file,
